+++ src/core/wi_table.h
@@ -43,7 +43,7 @@ wi_table_free(struct wi_table* table);
 bool
 wi_table_set(struct wi_table* table, wi_value key, wi_value value);
 void
-wi_table_set_foreign(struct wi_table* table, const char* name, wi_foreign_fn fn, int arity, bool is_variadic);
+wi_table_set_foreign(struct wi_table* table, const char* name, wi_foreign_fn fn, uint8_t arity, bool is_variadic);
 bool
 wi_table_get(struct wi_table* table, wi_value key, wi_value* value);
 bool
